internal/ast: use int

+++ internal/ast/ast.go
@@ -180,7 +180,7 @@ type (
 		stbase
 		Op         string
 		Arg        Expr  // Immediate argument for PUSH* / JUMP*.
-		Immediates []byte // Immediate arguments in brackets, e.g. dupn[x], exchange[x, y].
+		Immediates []int // Immediate arguments in brackets, e.g. dupn[x], exchange[x, y].
 		PushSize   byte  // For PUSH<n>, this is n+1.
 	}
